Extract Clear pipeline module
